Add ranks to prep sections
The prep section shows the ingredient groups that you should prep
before cooking. Before this commit they were ordered as they were
mentioned in the original recipe.

This commit grants a ranking. Prep groups should be ordered according
to the order in which they will be used in the recipe.